[NO-ISSUE] Improve test - Moved common methods into a separate go file - Using a single way to clean up resources with proper timeouts.
